Phase 2.6 step 0 (D-01, DR-013): widen ingestFlags to Uint32Array, add the explicit bit table
O13 is settled by D-01: the ingestion flag column keeps its positional encoding
(bit = rank in INGEST_FLAG_VALUES) but moves from Uint16Array to Uint32Array in
both copies of the frozen interface, which remain byte-for-byte identical. The
17th code of EX-DATA-45, MARKETPLACE_UNMAPPED - exactly the fallback ARB-60
creates for ADV-15 - was unstorable at bit 16 and could never be counted in
ingestFlagCounts (EX-DATA-46). 17 codes posted, 15 reserve bits.

src/types now owns the single source of the bit-to-code correspondence, as
DR-013 requires: INGEST_FLAG_BIT (a Readonly Record over the new IngestFlagCode
literal union), INGEST_FLAG_BIT_CAPACITY, and the helpers hasIngestFlag,
setIngestFlag and ingestFlagCodes. src/engine/flags.ts derives its four masks
from that table instead of recomputing an array index, and its 16-bit guard
becomes a capacity guard. No caller writes a literal shift any more.

Probes made green without touching their intent:
tests/review/D2/open-points.test.ts R-D2-18 and R-D2-18 (ADV-15).

Probes adapted (D-31, type adaptation forced by the widening, justified by
D-01): the two 16-bit thresholds those probes assert, the bitset16 expectation
on ingestFlags and the bitset count, plus the four review fixtures that
allocate the column themselves.

+++ src/engine/synthetic.ts
@@ -158,7 +158,7 @@ export function generateSyntheticDataset(options: SyntheticOptions): SyntheticDa
   const imageCount = new Uint8Array(n);
 
   const booleanFlags = new Uint16Array(n);
-  const ingestFlags = new Uint16Array(n);
+  const ingestFlags = new Uint32Array(n);
 
   // Prix « propre » avant injection (sert de base aux outliers).
   const cleanPrice = new Int32Array(n);
